+++ src/main/java/org/betonquest/betonquest/quest/event/kill/KillEvent.java
@@ -1,0 +1,24 @@
+package org.betonquest.betonquest.quest.event.kill;
+
+import org.betonquest.betonquest.api.profiles.Profile;
+import org.betonquest.betonquest.api.quest.event.Event;
+import org.betonquest.betonquest.exceptions.QuestRuntimeException;
+import org.bukkit.entity.Player;
+
+/**
+ * Kills the player.
+ */
+public class KillEvent implements Event {
+
+    /**
+     * Creates a new kill event.
+     */
+    public KillEvent() {
+    }
+
+    @Override
+    public void execute(final Profile profile) throws QuestRuntimeException {
+        final Player player = profile.getOnlineProfile().get().getPlayer();
+        player.setHealth(0);
+    }
+}