+++ src/main/java/org/betonquest/betonquest/quest/event/kill/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * {@link org.betonquest.betonquest.api.quest.event.Event Event} the implementation of the kill event.
+ */
+package org.betonquest.betonquest.quest.event.kill;