+++ src/main/java/org/betonquest/betonquest/quest/event/point/GlobalPointEventFactory.java
@@ -24,11 +24,10 @@ public class GlobalPointEventFactory implements EventFactory, StaticEventFactory
     }
 
     @Override
-    @SuppressWarnings("PMD.AvoidLiteralsInIfCondition")
     public Event parseEvent(final Instruction instruction) throws InstructionParseException {
         final String category = Utils.addPackage(instruction.getPackage(), instruction.next());
         final String number = instruction.next();
-        if (instruction.size() > 3) {
+        if (instruction.hasArgument("action")) {
             final String action = instruction.getOptional("action");
             if (action == null) {
                 throw new InstructionParseException("Missing modification action: " + instruction.current());